Adds Day 1.
I spent way too long trying to get the first and last digit out of one
regex. Once I split it into two, things went much more smoothly. I'm not
particularly happy with the dict strategy, but it worked on the first
try. There's probably a way to use the dict to create the regex, but I'm
not sure it'd be very clear.