Solves day 2 , part 1
Not knowing what is in part two, it seems prudent to load all the data
in a way that I can ask questions of it later. I don't work with SQL
often, but the shape of the data is easy enough to model, and now I
can query it with less hassle.