Adds day 3, part 1.
This was fun! Normally, thinking about doing spatial checks on text
makes me want to skip. My first instinct was to make a list of all the
numbers and symbols, and then iterate though all the numbers to check
each symbol in case they were close enough. I really didn't like the
performance implications of that approach, and luckily I realized I only
need to check the line above and below. It still took a leap to realize
that keeping only the previous numbers and symbols was sufficient.

I knew from working with regexes on day 1 that I could get the bounds
out of the match objects. That saved a ton of processing! Everything
else mostly came together smoothly, except I tried removing the matched
numbers from the current list immediately, which messed up the iterator.
Switching to the removal set fixed that quickly.

I am quite relieved that I got the math on the bounds checking correct
on the first try. I was very worried I'd have to think about that
more...